+++ src/modules/apartment.rs
@@ -8,7 +8,7 @@ pub struct Apartment {
     pub location_id: i32,
     pub location_level: i32,
     pub location_name: String,
-    pub size: f32,
+    pub size: f64,
     pub rooms: i32,
     pub price: String,
     pub additional_costs: i32,
